+++ server/src/core/index.js
@@ -1,9 +1,5 @@
-import copycat from './copycat';
-import ytcat from './ytsearchcat';
-import suggestbeat from './suggestbeat';
+import copycat from "./copycat";
+import ytcat from "./ytsearchcat";
+import suggestbeat from "./suggestbeat";
 
-export {
-    copycat,
-    ytcat,
-    suggestbeat,
-}
+export { copycat, ytcat, suggestbeat };
